Docs: upgrade to Create React App 4
