+++ .eslintrc.json
@@ -38,6 +38,7 @@
       }
     ],
     "import/extensions": ["error", "always", { "ignorePackages": true }],
+    "import/no-anonymous-default-export": "error",
     "import/no-extraneous-dependencies": "off",
     "import/no-namespace": "error",
     "import/no-relative-parent-imports": "error",
