+++ docs/package.json
@@ -24,7 +24,7 @@
     "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
-    "react-scripts": "3.4.0"
+    "react-scripts": "4.0.0"
   },
   "browserslist": [
     "last 2 versions",
