+++ package.json
@@ -26,7 +26,7 @@
     "@testing-library/react": "^11.0.4",
     "@testing-library/react-hooks": "^3.4.2",
     "babel-eslint": "10.1.0",
-    "babel-jest": "^26.5.2",
+    "babel-jest": "^26.6.0",
     "blueimp-md5": "^2.18.0",
     "caniuse-lite": "^1.0.30001146",
     "chalk": "^4.1.0",
@@ -35,7 +35,7 @@
     "cssnano": "^4.1.10",
     "cypress": "^5.3.0",
     "danger": "10.5.0",
-    "eslint": "7.10.0",
+    "eslint": "7.11.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-config-prettier": "^6.13.0",
     "eslint-plugin-cypress": "^2.11.2",
@@ -50,7 +50,7 @@
     "flow-bin": "^0.135.0",
     "husky": "^4.3.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.5.2",
+    "jest": "26.6.0",
     "jscodeshift": "^0.11.0",
     "lint-staged": "^10.4.0",
     "netlify-cli": "^2.65.1",
