+++ packages/gestalt/rollup.config.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line flowtype/require-valid-file-annotation
 import plugins from '../gestalt-core/build.js'; // eslint-disable-line import/no-relative-parent-imports
 
-export default {
+const rollupConfig = {
   input: 'src/index.js',
   output: [
     {
@@ -42,3 +42,5 @@ export default {
   ],
   plugins: plugins('gestalt'),
 };
+
+export default rollupConfig;
