+++ packages/gestalt/src/uniformRowLayout.js
@@ -10,7 +10,7 @@ const offscreen = (width, height = Infinity) => ({
   height,
 });
 
-export default <T>({
+const uniformRowLayout = <T>({
   cache,
   columnWidth = 236,
   gutter = 14,
@@ -68,3 +68,5 @@ export default <T>({
   }
   return positions;
 };
+
+export default uniformRowLayout;
